Rename app package in manifest
