+++ sample-app/src/main/kotlin/moxy/sample/app/AppActivity.kt
@@ -1,8 +1,7 @@
-package moxy.sample
+package moxy.sample.app
 
 import androidx.appcompat.app.AppCompatActivity
-import example.com.androidxsample.R
+import moxy.sample.R
 
 class AppActivity : AppCompatActivity(R.layout.activity_app) {
-
 }
